chore: merge tests
